Revert "added slides-from create/edit -- added toast-message -- added post and patch services in core/services/http.service.ts"
This reverts commit bfa45c832dcd019b5f10f285dedbea0decf90f97.
