+++ src/app/app.component.html
@@ -1,3 +1,1 @@
 <router-outlet></router-outlet>
-
-<p-toast key="toastMessage" position="bottom-center"></p-toast>
