+++ src/app/app.module.ts
@@ -3,8 +3,6 @@ import { CoreModule } from './core/core.module';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
-import { ToastModule } from 'primeng/toast';
-import { MessageService } from 'primeng/api';
 
 
 @NgModule({
@@ -14,12 +12,9 @@ import { MessageService } from 'primeng/api';
   imports: [
     BrowserModule,
     CoreModule,
-    FeaturesModule,
-    ToastModule
+    FeaturesModule
   ],
-  providers: [
-    MessageService
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
