+++ src/app/core/core.module.ts
@@ -6,8 +6,7 @@ import {
 import { HttpService } from "./services/http.service";
 
 @NgModule({
-  declarations: [
-  ],
+  declarations: [],
   imports: [CommonModule, HttpClientModule],
   providers: [HttpService],
 })
