+++ src/app/features/public/components/donations/thanks/thanks.component.scss
