+++ src/app/shared/primeng/primeng.module.ts
@@ -1,11 +1,10 @@
-import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
-
 import { ButtonModule } from "primeng/button";
 import { CalendarModule } from "primeng/calendar";
 import { CardModule } from 'primeng/card';
 import { CarouselModule } from "primeng/carousel";
 import { ChipsModule } from "primeng/chips";
+import { CommonModule } from "@angular/common";
 import { DialogModule } from "primeng/dialog";
 import { DividerModule } from "primeng/divider";
 import { DropdownModule } from "primeng/dropdown";
@@ -20,7 +19,6 @@ import { RadioButtonModule } from "primeng/radiobutton";
 import { RippleModule } from "primeng/ripple";
 import { SelectButtonModule } from "primeng/selectbutton";
 import { SidebarModule } from "primeng/sidebar";
-import { ToastModule } from 'primeng/toast';
 import { TooltipModule } from "primeng/tooltip";
 
 @NgModule({
@@ -47,7 +45,6 @@ import { TooltipModule } from "primeng/tooltip";
     RippleModule,
     SelectButtonModule,
     SidebarModule,
-    ToastModule,
     TooltipModule,
   ],
 })
