+++ src/styles.scss
@@ -5,9 +5,3 @@
 .w-50{
     width: 50%;
 }
-
-.text-danger {
-    color: red;
-    font-size: .8rem;
-    font-weight: 400;
-}
